requirements.txt
Django==1.4.10
PyJWT==0.1.2
South==0.7.6
amqp==1.3.3
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.13
celery==3.1.7
distribute==0.6.24
dj-database-url==0.2.2
django-appconf==0.6
django-celery==3.1.1
django-celery-with-redis==3.0
django-compressor==1.3
django-jsonify==0.2.1
django-mptt==0.6.0
django-polymorphic==0.5.3
django-redis==1.4.5
gunicorn==18.0
hiredis==0.1.1
httplib2==0.7.7
icalendar==3.2
kombu==3.0.8
mock==1.0.1
psycopg2==2.5.1
pytz==2013.9
redis==2.9.0
requests==0.14.2
six==1.5.1
twilio==3.4.1
wsgiref==0.1.2
django-heroku-postgresify==0.3
newrelic==2.8.0.7
django-storages==1.1.8
boto==2.22.0
amqplib==1.0.2
